thelastWar

+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: lnoaille <lnoaille@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/04/27 17:45:20 by lnoaille          #+#    #+#             */
-/*   Updated: 2020/05/10 02:38:18 by lnoaille         ###   ########.fr       */
+/*   Updated: 2020/05/13 18:17:00 by lnoaille         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	unsigned int pos;
+	size_t pos;
 
 	pos = 0;
 	if (n == 0)
